Rename methods to avoid ambiguity
 - `is_file_encrypted` -> `is_filtered_file`
 - `find_encrypted_files` -> `find_filtered_files`

Indeed, the intent of those methods is to look for the files which have the git filter applied to them, rather than looking if the file's content looks like encrypted data or not.
The files might have their content currently decrypted on disk in the working copy (because they have been decrypted by the smudge filter on checkout), yet what we want to know with these methods is what files are managed/processed by the filter, so we know which files to encrypt/decrypt during `lock`/`unlock` commands, and know which files to print during `status` command.